requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
akshare
google-generativeai
openai
requests
schedule
tenacity
python-multipart
PyYAML
lark-oapi
tushare
google-search-results
tavily
numpy
efinance
